feat: switch solver arrays from f64 to f32 to halve worker memory
Calcium imaging data has far more noise than f32's ~1e-7 precision
limit, so single precision is sufficient for signal arrays. This
halves per-worker solver buffer memory (~2MB → ~1MB for typical
traces), which matters with N parallel workers.

Rust: Vec<f64> → Vec<f32> for all signal arrays (trace, solution,
kernel, etc). Scalar FISTA state (t_fista, prev_objective, tolerance,
lipschitz_constant) stays f64 for convergence tracking precision.
Kernel built in f64 then cast to f32. Objective function accumulates
in f64. export_state/load_state use 4 bytes per element.

TypeScript: solver output types Float64Array → Float32Array. Input
traces stay Float64Array (raw .npy data); pool-worker converts at
the WASM boundary. residualTrace stays Float64Array (f64 - f32
auto-promotes in JS). downsample and metrics accept both array types.

+++ src/lib/multi-cell-store.ts
@@ -12,8 +12,8 @@ export type SelectionMode = 'top-active' | 'random' | 'manual';
 export interface CellTraces {
   cellIndex: number;
   raw: Float64Array;
-  deconvolved: Float64Array;
-  reconvolution: Float64Array;
+  deconvolved: Float32Array;
+  reconvolution: Float32Array;
   windowStartSample?: number;
 }
 
@@ -45,8 +45,8 @@ function updateOneCellStatus(cellIndex: number, status: CellSolverStatus): void 
 
 function updateOneCellTraces(
   cellIndex: number,
-  deconvolved: Float64Array,
-  reconvolution: Float64Array,
+  deconvolved: Float32Array,
+  reconvolution: Float32Array,
   windowStartSample?: number,
 ): void {
   setMultiCellResults(prev => {
